Added Leave Details

+++ app/Http/Controllers/LeaveController.php
@@ -6,6 +6,8 @@ use App\Leave;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use Carbon\Carbon;
+use App\User;
+use App\LeaveType;
 
 class LeaveController extends Controller
 {
@@ -24,7 +26,7 @@ class LeaveController extends Controller
     }
     public function getAllLeaves()
     {
-        return DataTables::of(Leave::with('user')->select('leaves.*'))
+        return DataTables::of(Leave::with('user')->with('admin')->with('type')->select('leaves.*'))
         ->addColumn('action', function ($leave) {
             return '<div class="d-flex align-items-baseline">
                         <a href="'.route('leave.show',$leave->id).'" class="btn btn-sm btn-rounded bg-white p-0 m-0 pr-2" data-toggle="tooltip" data-placement="top" title="Show Department Details">
@@ -44,6 +46,12 @@ class LeaveController extends Controller
         ->addColumn('user.full_name', function($leave){
             return $leave->user->full_name;
         })
+        ->addColumn('type.name', function($leave){
+            return $leave->type->name;
+        })
+        ->addColumn('admin.full_name', function($leave){
+            return $leave->admin->full_name;
+        })
         ->editColumn('created_at', function($leave){
             return $leave->created_at->format('F d, Y');
         })
@@ -59,7 +67,10 @@ class LeaveController extends Controller
      */
     public function create()
     {
-        return view('leave.create');
+        $this->authorize('create', Leave::class);
+        $leaveTypes = LeaveType::all();
+        $users = User::all();
+        return view('leave.create', compact('users', 'leaveTypes'));
     }
 
     /**
@@ -71,13 +82,13 @@ class LeaveController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'type' => 'required|string|max:255',
+            'leave_type_id' => 'required|string|max:255',
             'reason' => 'required|string|max:255',
             'start_date' => 'required|string|max:20',
-            'end_date' => 'required|string|max:20',
+            'end_date' => 'required|string|max:20'
         ]);
         /**Check if end date is greater than start date */
-        if (Carbon::parse($request->end_date) > Carbon::parse($request->start_date)) {
+        if (Carbon::parse($request->end_date) >= Carbon::parse($request->start_date)) {
             /**Check if end date and start date is greater than the date today  */
             if (Carbon::parse($request->end_date) < Carbon::now() || Carbon::parse($request->start_date) < Carbon::now()) {
                 alert()->error('Error!','Leave date must be greater than today')->showConfirmButton('Confirm', '#3085d6');;
@@ -85,28 +96,11 @@ class LeaveController extends Controller
                 return redirect()->back()->withInput();
             }
             else{
-                /**Check if the leave type is Vacation to verify if the leave is filed 3 days before it will start */
-                if ($request->type == 'Vacation') {
-                    if (Carbon::parse($request->start_date) >= Carbon::now()->addDays(2)) {
-                        Leave::create($request->all());
+                Leave::create($request->all());
 
-                        toast('Leave successfully filed!', 'success', 'top');
+                toast('Leave successfully filed!', 'success', 'top');
 
-                        return redirect()->route('leave.index');
-                    }
-                    else{
-                        alert()->error('Error!','Leave must be filed 3 days before the start date!')->showConfirmButton('Confirm', '#3085d6');;
-
-                        return redirect()->back()->withInput();
-                    }
-                }
-                else{
-                        Leave::create($request->all());
-
-                        toast('Leave successfully filed!', 'success', 'top');
-
-                        return redirect()->route('leave.index');
-                }
+                return redirect()->route('leave.index');
             }
         }
         else{
@@ -135,7 +129,9 @@ class LeaveController extends Controller
      */
     public function edit(Leave $leave)
     {
-        return view('leave.edit', compact('leave'));
+        $users = User::all();
+        $leaveTypes = LeaveType::all();
+        return view('leave.edit', compact('leave', 'users', 'leaveTypes'));
     }
 
     /**
@@ -147,15 +143,15 @@ class LeaveController extends Controller
      */
     public function update(Request $request, Leave $leave)
     {
-        // $this->authorize('update', $leave);
+        $this->authorize('update', $leave);
         $request->validate([
-            'type' => 'required|string|max:255',
+            'leave_type_id' => 'required|string|max:255',
             'reason' => 'required|string|max:255',
             'start_date' => 'required|string|max:20',
             'end_date' => 'required|string|max:20',
         ]);
         /**Check if end date is greater than start date */
-        if (Carbon::parse($request->end_date) > Carbon::parse($request->start_date)) {
+        if (Carbon::parse($request->end_date) >= Carbon::parse($request->start_date)) {
             /**Check if end date and start date is greater than the date today  */
             if (Carbon::parse($request->end_date) < Carbon::now() || Carbon::parse($request->start_date) < Carbon::now()) {
                 alert()->error('Error!','Leave date must be greater than today')->showConfirmButton('Confirm', '#3085d6');;
@@ -163,28 +159,11 @@ class LeaveController extends Controller
                 return redirect()->back()->withInput();
             }
             else{
-                /**Check if the leave type is Vacation to verify if the leave is filed 3 days before it will start */
-                if ($request->type == 'Vacation') {
-                    if (Carbon::parse($request->start_date) >= Carbon::now()->addDays(2)) {
-                        $leave->update($request->all());
+                $leave->update($request->all());
 
-                        toast('Leave successfully filed!', 'success', 'top');
+                toast('Leave has been successfully updated!', 'success', 'top');
 
-                        return redirect()->route('leave.index');
-                    }
-                    else{
-                        alert()->error('Error!','Leave must be filed 3 days before the start date!')->showConfirmButton('Confirm', '#3085d6');;
-
-                        return redirect()->back()->withInput();
-                    }
-                }
-                else{
-                        $leave->update($request->all());
-
-                        toast('Leave successfully filed!', 'success', 'top');
-
-                        return redirect()->route('leave.index');
-                }
+                return redirect()->route('leave.index');
             }
         }
         else{
@@ -202,7 +181,7 @@ class LeaveController extends Controller
      */
     public function destroy(Leave $leave)
     {
-        // $this->authorize('delete', $leave);
+        $this->authorize('delete', $leave);
         $leave->delete();
     }
 }
